feat: implement logs widget
